+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* update Toolbar to ToolbarGroup and ToolbarButton
 
 #### 2.1.2 / 2020-09-15
 * use same _block name_ for oembed